requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.1.14
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
environs==9.5.0
executing==0.8.3
idna==3.3
inflection==0.5.1
ipython==8.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5

